Define per-cluster server and client targets in GN
Have chip_data_model targets depend on these, instead of making up a source set on the fly.
