+++ examples/network-manager-app/linux/main.cpp
@@ -16,8 +16,8 @@
  */
 
 #include <AppMain.h>
-#include <app/clusters/thread-network-directory-server/thread-network-directory-server.h>
-#include <app/clusters/wifi-network-management-server/wifi-network-management-server.h>
+#include <app/clusters/thread-network-directory/thread-network-directory-server.h>
+#include <app/clusters/wifi-network-management/wifi-network-management-server.h>
 #include <lib/core/CHIPSafeCasts.h>
 #include <lib/support/CodeUtils.h>
 #include <lib/support/Span.h>
