+++ src/app/chip_data_model.gni
@@ -247,7 +247,12 @@ template("chip_data_model") {
       defines = []
     }
 
-    foreach(cluster, _cluster_sources) {
+    foreach(cluster_target, _cluster_sources) {
+      # Work out the legacy ${cluster} directory name
+      cluster_target_parts = []
+      cluster_target_parts = string_split(cluster_target, ":")
+      cluster = cluster_target_parts[0]
+
       if (cluster == "door-lock-server") {
         sources += [
           "${_app_root}/clusters/${cluster}/door-lock-server-callback.cpp",
@@ -299,14 +304,6 @@ template("chip_data_model") {
         # TODO: DefaultOTARequestor depends on controller, however the cluster server itself does not.
         # Maybe DefaultOTARequestor and related code should have its own source set.
         deps += [ "${chip_root}/src/controller:interactions" ]
-      } else if (cluster == "bindings") {
-        sources += [
-          "${_app_root}/clusters/${cluster}/${cluster}.cpp",
-          "${_app_root}/clusters/${cluster}/BindingManager.cpp",
-          "${_app_root}/clusters/${cluster}/BindingManager.h",
-          "${_app_root}/clusters/${cluster}/PendingNotificationMap.cpp",
-          "${_app_root}/clusters/${cluster}/PendingNotificationMap.h",
-        ]
       } else if (cluster == "time-synchronization-server") {
         sources += [
           "${_app_root}/clusters/${cluster}/${cluster}.cpp",
@@ -428,14 +425,6 @@ template("chip_data_model") {
           "${_app_root}/clusters/${cluster}/${cluster}.h",
           "${_app_root}/clusters/${cluster}/WaterHeaterManagementTestEventTriggerHandler.h",
         ]
-      } else if (cluster == "thread-network-directory-server") {
-        sources += [
-          "${_app_root}/clusters/${cluster}/${cluster}.cpp",
-          "${_app_root}/clusters/${cluster}/${cluster}.h",
-          "${_app_root}/clusters/${cluster}/DefaultThreadNetworkDirectoryStorage.cpp",
-          "${_app_root}/clusters/${cluster}/DefaultThreadNetworkDirectoryStorage.h",
-          "${_app_root}/clusters/${cluster}/ThreadNetworkDirectoryStorage.h",
-        ]
       } else if (cluster == "thermostat-server") {
         sources += [
           "${_app_root}/clusters/${cluster}/${cluster}-atomic.cpp",
@@ -452,8 +441,68 @@ template("chip_data_model") {
           "${_app_root}/clusters/${cluster}/ArlEncoder.cpp",
           "${_app_root}/clusters/${cluster}/ArlEncoder.h",
         ]
-      } else {
+      } else if (cluster == "account-login-server" ||
+                 cluster == "administrator-commissioning-server" ||
+                 cluster == "application-basic-server" ||
+                 cluster == "audio-output-server" ||
+                 cluster == "basic-information" ||
+                 cluster == "bridged-device-basic-information-server" ||
+                 cluster == "channel-server" || cluster == "chime-server" ||
+                 cluster == "color-control-server" ||
+                 cluster == "commissioner-control-server" ||
+                 cluster == "content-app-observer" ||
+                 cluster == "content-control-server" ||
+                 cluster == "content-launch-server" ||
+                 cluster == "ecosystem-information-server" ||
+                 cluster == "electrical-power-measurement-server" ||
+                 cluster == "energy-preference-server" ||
+                 cluster == "ethernet-network-diagnostics-server" ||
+                 cluster == "fan-control-server" ||
+                 cluster == "fault-injection-server" ||
+                 cluster == "fixed-label-server" ||
+                 cluster == "general-commissioning-server" ||
+                 cluster == "general-diagnostics-server" ||
+                 cluster == "group-key-mgmt-server" ||
+                 cluster == "groups-server" || cluster == "identify-server" ||
+                 cluster == "keypad-input-server" ||
+                 cluster == "laundry-dryer-controls-server" ||
+                 cluster == "laundry-washer-controls-server" ||
+                 cluster == "level-control" ||
+                 cluster == "localization-configuration-server" ||
+                 cluster == "low-power-server" ||
+                 cluster == "media-input-server" ||
+                 cluster == "media-playback-server" ||
+                 cluster == "messages-server" ||
+                 cluster == "microwave-oven-control-server" ||
+                 cluster == "network-commissioning" ||
+                 cluster == "occupancy-sensor-server" ||
+                 cluster == "on-off-server" ||
+                 cluster == "operational-credentials-server" ||
+                 cluster == "ota-provider" ||
+                 cluster == "power-source-configuration-server" ||
+                 cluster == "power-source-server" ||
+                 cluster == "power-topology-server" ||
+                 cluster == "pump-configuration-and-control-server" ||
+                 cluster == "pump-configuration-and-control-client" ||
+                 cluster == "refrigerator-alarm-server" ||
+                 cluster == "sample-mei-server" ||
+                 cluster == "software-diagnostics-server" ||
+                 cluster == "switch-server" ||
+                 cluster == "target-navigator-server" ||
+                 cluster == "test-cluster-server" ||
+                 cluster == "thermostat-user-interface-configuration-server" ||
+                 cluster == "thermostat-client" ||
+                 cluster == "time-format-localization-server" ||
+                 cluster == "timer-server" || cluster == "user-label-server" ||
+                 cluster == "valve-configuration-and-control-server" ||
+                 cluster == "wake-on-lan-server" ||
+                 cluster == "wifi-network-diagnostics-server" ||
+                 cluster == "window-covering-server") {
+        # Fallback for clusters not yet migrated, do not add to this list
         sources += [ "${_app_root}/clusters/${cluster}/${cluster}.cpp" ]
+      } else {
+        # New clusters should follow this pattern and provide a BUILD.gn with a "server" or "client" target as appropriate
+        public_deps += [ "${_app_root}/clusters/${cluster_target}" ]
       }
     }
 
@@ -470,6 +519,7 @@ template("chip_data_model") {
       "${chip_root}/src/app/common:cluster-objects",
       "${chip_root}/src/app/common:enums",
       "${chip_root}/src/app/server",
+      "${chip_root}/src/app/util:binding-table-headers",
       "${chip_root}/src/app/util:types",
       "${chip_root}/src/app/util/persistence",
       "${chip_root}/src/lib/core",
