+++ src/app/clusters/bindings/BUILD.gn
@@ -1,0 +1,31 @@
+# Copyright (c) 2024 Project CHIP Authors
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+import("//build_overrides/chip.gni")
+
+source_set("server") {
+  deps = [
+    "${chip_root}/src/app",
+    "${chip_root}/src/app/server",
+    "${chip_root}/src/app/util:binding-table-headers",
+  ]
+  sources = [
+    "BindingManager.cpp",
+    "BindingManager.h",
+    "PendingNotificationMap.cpp",
+    "PendingNotificationMap.h",
+    "bindings.cpp",
+    "bindings.h",
+  ]
+}
