+++ src/app/clusters/thread-network-directory/thread-network-directory-server.h
@@ -20,8 +20,8 @@
 #include <app-common/zap-generated/cluster-objects.h>
 #include <app/AttributeAccessInterface.h>
 #include <app/CommandHandlerInterface.h>
-#include <app/clusters/thread-network-directory-server/DefaultThreadNetworkDirectoryStorage.h>
-#include <app/clusters/thread-network-directory-server/ThreadNetworkDirectoryStorage.h>
+#include <app/clusters/thread-network-directory/DefaultThreadNetworkDirectoryStorage.h>
+#include <app/clusters/thread-network-directory/ThreadNetworkDirectoryStorage.h>
 #include <app/server/Server.h>
 #include <lib/core/CHIPError.h>
