+++ src/app/tests/TestDefaultThreadNetworkDirectoryStorage.cpp
@@ -15,7 +15,7 @@
  *    limitations under the License.
  */
 
-#include <app/clusters/thread-network-directory-server/DefaultThreadNetworkDirectoryStorage.h>
+#include <app/clusters/thread-network-directory/DefaultThreadNetworkDirectoryStorage.h>
 #include <lib/core/CHIPPersistentStorageDelegate.h>
 #include <lib/support/DefaultStorageKeyAllocator.h>
 #include <lib/support/TestPersistentStorageDelegate.h>
